aftermath to ee3b0f41d1825c11b28da2e8bc29107d04088f6e: fixed double precision for pairing list variance/stddev

+++ java/com.sap.sse/src/com/sap/sse/pairinglist/impl/PairingListTemplateImpl.java
@@ -826,7 +826,7 @@ public class PairingListTemplateImpl implements PairingListTemplate {
             }
         }
         // expression in Math.sqrt() is equal to variance / n
-        standardDev = Math.sqrt((exp2 - exp*exp / n) / n);
+        standardDev = Math.sqrt((exp2 - (double) (exp*exp) / n) / n);
         return standardDev;
     }
